requirements.txt
uvicorn
fastapi
typing
pydantic
mysql-connector-python
pymongo
dnspython
mysqlclient
